requirements.txt
asttokens==2.4.1
autopep8==1.4.4
blinker==1.4
certifi==2023.11.17
charset-normalizer==3.3.2
Click==7.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-Testing==0.8.1
idna==3.6
ipython==8.18.1
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.10.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.5.0
Pygments==2.17.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==0.16.0
